Sagi

+++ app/src/main/java/com/example/youtube/MainActivity.java
@@ -1,10 +1,12 @@
 package com.example.youtube;
 
 import android.annotation.SuppressLint;
+import android.content.Context;
 import android.content.Intent;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.os.Bundle;
 import android.view.MenuInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.ImageView;
@@ -26,6 +28,7 @@ import com.example.youtube.databinding.ActivityMainBinding;
 import com.example.youtube.entities.User;
 import com.example.youtube.repositories.UserRepository;
 import com.example.youtube.viewModels.VideoViewModel;
+import com.example.youtube.viewModels.VideoViewModelFactory;
 
 import java.util.List;
 
@@ -34,6 +37,7 @@ public class MainActivity extends AppCompatActivity {
     private ActivityMainBinding binding;
     private VideosListAdapter videoAdapter;
     private UsersListAdapter userAdapter;
+    private VideoViewModel videoViewModel;
     private ImageView youBtn;
     private UserRepository userRepository;
     private User loggedInUser;
@@ -48,26 +52,18 @@ public class MainActivity extends AppCompatActivity {
         userRepository = UserRepository.getInstance(this);
 
         ImageButton btnToggleDark = binding.modeBtn;
-        btnToggleDark.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                boolean isDarkMode = (AppCompatDelegate.getDefaultNightMode() == AppCompatDelegate.MODE_NIGHT_YES);
-                if (isDarkMode) {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-                } else {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-                }
-                recreate();
-            }
+        btnToggleDark.setOnClickListener(view -> {
+            boolean isDarkMode = (AppCompatDelegate.getDefaultNightMode() == AppCompatDelegate.MODE_NIGHT_YES);
+            AppCompatDelegate.setDefaultNightMode(isDarkMode ? AppCompatDelegate.MODE_NIGHT_NO : AppCompatDelegate.MODE_NIGHT_YES);
+            recreate();
         });
 
         RecyclerView lstVideos = binding.lstVideos;
         videoAdapter = new VideosListAdapter(this);
 
-        VideoViewModel viewModel = new ViewModelProvider(this).get(VideoViewModel.class);
-        viewModel.getVideos().observe(this, videos -> {
-            videoAdapter.setVideos(videos);
-        });
+        VideoViewModelFactory factory = new VideoViewModelFactory(getApplication());
+        videoViewModel = new ViewModelProvider(this, factory).get(VideoViewModel.class);
+        videoViewModel.getVideos().observe(this, videos -> videoAdapter.setVideos(videos));
 
         lstVideos.setAdapter(videoAdapter);
         lstVideos.setLayoutManager(new LinearLayoutManager(this));
@@ -94,7 +90,11 @@ public class MainActivity extends AppCompatActivity {
 
         ImageButton btnUploadVideo = binding.uploadBtn;
         btnUploadVideo.setOnClickListener(v -> {
-            if (userRepository.getLoggedInUser() == null) {
+            if (isOffline()) {
+                Toast.makeText(this, "you are offline, to upload a video please connect to the internet first", Toast.LENGTH_SHORT).show();
+                return;
+            }
+            if (loggedInUser == null) {
                 Toast.makeText(MainActivity.this, "You need to be logged in to upload a video", Toast.LENGTH_SHORT).show();
                 return;
             }
@@ -106,8 +106,7 @@ public class MainActivity extends AppCompatActivity {
         youBtn = binding.youBtn;
         if (loggedInUser == null) {
             setLoggedOutState();
-        }
-        else {
+        } else {
             setLoggedInState();
         }
     }
@@ -115,6 +114,10 @@ public class MainActivity extends AppCompatActivity {
     @Override
     protected void onResume() {
         super.onResume();
+        RecyclerView lstVideos = binding.lstVideos;
+        videoViewModel.getVideos().observe(this, videos -> videoAdapter.setVideos(videos));
+        lstVideos.setAdapter(videoAdapter);
+        lstVideos.setLayoutManager(new LinearLayoutManager(this));
         LiveData<List<User>> users = userRepository.getAllUsers();
         users.observe(this, userList -> userAdapter.setUsers(userList));
     }
@@ -131,7 +134,7 @@ public class MainActivity extends AppCompatActivity {
                         .into(youBtn);
             }
         }
-        youBtn.setOnClickListener(v -> showUserOptionsMenu(v));
+        youBtn.setOnClickListener(this::showUserOptionsMenu);
     }
 
     private void showUserOptionsMenu(View view) {
@@ -139,29 +142,35 @@ public class MainActivity extends AppCompatActivity {
         MenuInflater inflater = popupMenu.getMenuInflater();
         inflater.inflate(R.menu.user_options_menu, popupMenu.getMenu());
 
-        popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-            @Override
-            public boolean onMenuItemClick(MenuItem item) {
-                CharSequence title = item.getTitle();
-                assert title != null;
+        popupMenu.setOnMenuItemClickListener(item -> {
+            CharSequence title = item.getTitle();
+            if (title != null) {
                 if (title.equals("Log Out")) {
                     userRepository.logoutUser();
                     setLoggedOutState();
                     return true;
                 } else if (title.equals("Edit User")) {
+                    if (isOffline()) {
+                        Toast.makeText(this, "you are offline, to edit a user please connect to the internet first", Toast.LENGTH_SHORT).show();
+                        return true;
+                    }
                     Intent editIntent = new Intent(MainActivity.this, EditUserActivity.class);
                     startActivity(editIntent);
                     return true;
                 } else if (title.equals("Delete User")) {
                     if (loggedInUser != null) {
+                        if (isOffline()) {
+                            Toast.makeText(this, "you are offline, to delete a user please connect to the internet first", Toast.LENGTH_SHORT).show();
+                            return true;
+                        }
                         userRepository.delete(loggedInUser.getUsername(), loggedInUser.getToken());
                         userRepository.logoutUser();
                         setLoggedOutState();
                     }
                     return true;
                 }
-                return false;
             }
+            return false;
         });
 
         popupMenu.show();
@@ -175,5 +184,10 @@ public class MainActivity extends AppCompatActivity {
             Intent i = new Intent(MainActivity.this, LogInActivity.class);
             startActivity(i);
         });
+    }
+    private boolean isOffline() {
+        ConnectivityManager connectivityManager = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo networkInfo = connectivityManager.getActiveNetworkInfo();
+        return networkInfo == null || !networkInfo.isConnected();
     }
 }
